Correcting conditions on trigger window parameters
The first problem was the use of uin16_t. It cannot store values below
-32767. Also the result of unsigned + signed numbers gives a unsigned
number which led to an uncorrect if condition.

Trigger window parameters were partly invalid in CAEN documentation
(in rev. 15 and not totally corrected in rev. 16). It's better to
check CERN HPTDC documentation.

Trigger time substraction was also enabled to avoid counter reset during
an event.